Implement video downloading

+++ Sources/App/Helpers/ShellUtil.swift
@@ -46,9 +46,11 @@ enum ShellUtil {
 
     // MARK: Private properties
 
+    private static var useVirtualEnvironment = Environment.get("USE_PYTHON_VIRTUAL_ENV") == String(true)
     private static var environmentActivationCommand: String {
         "source \(Constant.defaultEnvironmentPath)/bin/activate"
     }
+
 }
 
 // MARK: - Internal methods
@@ -60,21 +62,22 @@ extension ShellUtil {
             let output = try shellOut(to: "whereis \(launchBinary.location)")
 
             // TODO: check
-            if output.components(separatedBy: " ").count < 2 {
-                Logger.shellLogger.error("Launch binary not found: \(launchBinary.location)")
-                throw ShellUtilError.missingLaunchBinary(launchBinary.location)
-            }
+//            if output.components(separatedBy: " ").count < 2 {
+//                Logger.shellLogger.error("Launch binary not found: \(launchBinary.location)")
+//                throw ShellUtilError.missingLaunchBinary(launchBinary.location)
+//            }
         }
     }
 
-    static func createVirtualEnvironment() throws {
+    static func createVirtualEnvironmentIfNeeded() throws {
+        guard useVirtualEnvironment else { return }
         do {
-            try checkVirtualEnvironment()
+            try checkVirtualEnvironmentIfPossible()
         } catch {
             do {
                 Logger.shellLogger.info("Creating virtual python environment")
                 try shellOut(to: "\(LaunchBinary.python.location) -m venv \(Constant.defaultEnvironmentPath)")
-                try checkVirtualEnvironment()
+                try checkVirtualEnvironmentIfPossible()
             } catch {
                 Logger.shellLogger.error("Cannot reate virtual python environment")
                 try shellOut(to: "rm -rf \(Constant.defaultEnvironmentPath)")
@@ -83,7 +86,8 @@ extension ShellUtil {
         }
     }
 
-    static func installDownloader() throws {
+    static func installDownloaderIfNeeded() throws {
+        guard useVirtualEnvironment else { return }
         Logger.shellLogger.info("Installing downloader")
         try shellOut(to: [
             environmentActivationCommand,
@@ -93,7 +97,7 @@ extension ShellUtil {
     }
 
     @discardableResult
-    static func downloadVideo(url: String, with arguments: [String], useVirtualEnvironment: Bool) throws -> String {
+    static func downloadVideo(url: String, with arguments: [String]) throws -> String {
         var commands = ["youtube-dl \(arguments.map { $0 }.joined(separator: " ")) \(url)"]
         if useVirtualEnvironment {
             commands.insert(environmentActivationCommand, at: .zero)
@@ -107,7 +111,8 @@ extension ShellUtil {
 // MARK: - Helpers
 
 extension ShellUtil {
-    private static func checkVirtualEnvironment() throws {
+    private static func checkVirtualEnvironmentIfPossible() throws {
+        guard useVirtualEnvironment else { return }
         Logger.shellLogger.info("Checking virtual python environment")
         try shellOut(to: [
             environmentActivationCommand,
